Move follow DTO into response subfolder
