+++ src/modules/follow/dto/response/follow.dto.ts
@@ -1,5 +1,5 @@
 import { Exclude, Expose, Type } from 'class-transformer';
-import { UserDto } from 'src/modules/auth/dto/user.dto';
+import { UserDto } from 'src/modules/auth/dto/response/user.dto';
 
 
 export class FollowDto {
